create separate entrypoints for shared package
