+++ packages/shared/data/index.ts
@@ -1,0 +1,2 @@
+export { links } from "./breadchain";
+export { memberProjects } from "./projects";
