--- packages/shared/src/Data/index.ts
@@ -1,2 +1,0 @@
-export * from "./ProjectData";
-export * from "./Links";
